Refactor: Collapse duplicate merge wait and plan aliases #1264
Share one merge-in-progress detector and drop identity aliases so the deal merge plan has a single ID and attribute contract.
